fix: raise error on unsupported hostname change
Error out if hostname changed. Also add some testing.
